move KVCacheSpecKind to kvevents pkg
